Reports validate configuration data

+++ src/main/java/es/eucm/gleaner/analysis/Main.java
@@ -35,31 +35,70 @@ public class Main {
 		DBObject versionData = db.getCollection("versions").findOne(
 				new BasicDBObject("_id", new ObjectId(versionId)));
 
-		Configuration config = new Configuration();
-		config.set("mongo.input.uri", "mongodb://" + mongoHost + ":"
-				+ mongoPort + "/" + mongoDB + ".traces_" + versionId);
+		Object loading = versionData.containsField("analyzing") ? versionData
+				.get("analyzing") : Boolean.FALSE;
 
-		// Collection of traces<null, trace>
-		JavaPairRDD<Object, BSONObject> traces = sc.newAPIHadoopRDD(config,
-				MongoInputFormat.class, Object.class, BSONObject.class);
+		if (loading instanceof Boolean && !(Boolean) loading) {
+			// JavaSparkContext sc = new JavaSparkContext();
 
-		GameplaysAnalysis gameplaysAnalysis = new GameplaysAnalysis();
-		gameplaysAnalysis.read(versionData);
-		JavaPairRDD<Object, BSONObject> gameplaysResults = gameplaysAnalysis
-				.calculateGameplayResults(traces);
+			db.getCollection("versions").update(
+					new BasicDBObject("_id", versionData.get("_id")),
+					new BasicDBObject("$set",
+							new BasicDBObject("loading", true)));
 
-        // Write to mongo gameplays
-        db.getCollection("gameplaysresults_" + versionId).drop();
+			Configuration config = new Configuration();
+			config.set("mongo.input.uri", "mongodb://" + mongoHost + ":"
+					+ mongoPort + "/" + mongoDB + ".traces_" + versionId);
 
-        config.set("mongo.output.uri", "mongodb://" + mongoHost + ":"
-                + mongoPort + "/" + mongoDB + ".gameplaysresults_" + versionId);
-        gameplaysResults.saveAsNewAPIHadoopFile("file:///bogus", Object.class,
-                Object.class, MongoOutputFormat.class, config);
+			// Collection of traces<null, trace>
+			JavaPairRDD<Object, BSONObject> traces = sc.newAPIHadoopRDD(config,
+					MongoInputFormat.class, Object.class, BSONObject.class);
 
-        ArrayList<DBObject> segmentResults = gameplaysAnalysis
-				.calculateSegments(gameplaysResults, versionData);
-        // Write to mongo segment results
-        db.getCollection("segments_" + versionId).drop();
-        db.getCollection("segments_" + versionId).insert(segmentResults);
+			GameplaysAnalysis gameplaysAnalysis = new GameplaysAnalysis();
+			gameplaysAnalysis.read(versionData);
+			JavaPairRDD<Object, BSONObject> gameplaysResults = gameplaysAnalysis
+					.calculateGameplayResults(traces);
+
+			// Write to mongo gameplays
+			db.getCollection("gameplaysresults_" + versionId).drop();
+
+			config.set("mongo.output.uri", "mongodb://" + mongoHost + ":"
+					+ mongoPort + "/" + mongoDB + ".gameplaysresults_"
+					+ versionId);
+			gameplaysResults
+					.saveAsNewAPIHadoopFile("file:///bogus", Object.class,
+							Object.class, MongoOutputFormat.class, config);
+
+			// Read players
+
+					BSONObject.class);
+
+			gameplaysResults = gameplaysResults
+					.mapToPair(new ExtractFieldAsKey("gameplayId"))
+					.join(gameplays.mapToPair(new ExtractFieldAsKey("_id")))
+					.mapToPair(
+							new PairFunction<Tuple2<String, Tuple2<BSONObject, BSONObject>>, Object, BSONObject>() {
+								@Override
+								public Tuple2<Object, BSONObject> call(
+										Tuple2<String, Tuple2<BSONObject, BSONObject>> tuple2)
+										throws Exception {
+									BSONObject result = tuple2._2._1;
+									BSONObject gameplay = tuple2._2._2;
+									return new Tuple2<Object, BSONObject>(Q
+											.get("playerId", gameplay)
+											.toString(), result);
+								}
+							});
+
+			ArrayList<DBObject> segmentResults = gameplaysAnalysis
+					.calculateSegments(gameplaysResults, versionData);
+			// Write to mongo segment results
+			db.getCollection("segments_" + versionId).drop();
+			db.getCollection("segments_" + versionId).insert(segmentResults);
+			db.getCollection("versions").update(
+					new BasicDBObject("_id", versionData.get("_id")),
+					new BasicDBObject("$set", new BasicDBObject("analyzing",
+							false)));
+		}
 	}
 }
